Extracting the mounting of the app into a function for reuse in acceptance tests.
This allows to

    Pass a fresh instance of the router per each test running in parallel without interfering with each other
    Use a real router instead of a mocked one.
